Fix race condition in GelfLoggingFilter$CountingHttpServletResponseWrapper
GelfLoggingFilter had a race condition when the underlying OutputStream of the
HttpServletResponse had been reset or closed by another Filter (or Jetty Handler).

Instead of eagerly initializing the OutputStream in the constructor of
CountingHttpServletResponseWrapper we now lazily initialize it in the getOutputStream()
method and also handle resets with #reset() and #resetBuffer() correctly.