Add a new UncaughtExceptionHandler that inits a GELF appender before logging
Since Dropwizard 0.8, the logging system is shut down when the run method of
a ConfiguredCommand is left (see
https://github.com/dropwizard/dropwizard/commit/cebfb632b71e3524a678e7c1f1df64264de7dcae)
Unfortunately, that also means that the GELF appender we bootstraped before
entering the run method is stopped as well. The new exception handler
(re-)adds a GELF appender before doing the actual logging call.