Upgrade to Dropwizard 1.2.0
Upgrade to the latest version of Dropwizard and remove unnecessary
field `threshold` from `GelfAppenderFactory`.  It's already present
in `AbstractAppenderFactory`, so the `threshold` field in
`GelfAppenderFactory` just shadowed it. Unfortunately in 1.2.0, the type
of public getters of the `threshold` field in `AbstractAppenderFactory`
changed their signature (they are now serialized and deserialized as
strings), so `GelfAppenderFactory` didn't compile. But because this
field is superfluous, we can safely remove it.